servers/github/server.py
"""GitHub MCP Server - Access GitHub repositories through MCP.

This server will be fully implemented in Phase 6.
"""

import asyncio

import yaml

# Placeholder for Phase 6 implementation


def load_config(config_path: str = "config.yaml") -> dict:
    """Load server configuration."""
    with open(config_path) as f:
        return yaml.safe_load(f)


async def main():
    """Main entry point for the GitHub MCP server."""
    print("GitHub MCP Server - Phase 6 Implementation Pending")
    print("Run with: mcp-inspector python server.py")


if __name__ == "__main__":
    asyncio.run(main())
