pyproject.toml
[project]
name = "axiom-mcp-servers"
version = "0.1.0"
description = "MCP server implementations for Axiom - filesystem and GitHub data sources"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"

dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0.2",
    "aiofiles>=24.1.0",

    # GitHub server
    "pygithub>=2.5.0",
    "gitpython>=3.1.43",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

gdrive = [
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.1",
    "google-api-python-client>=2.149.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
